moved Inteface and two other classes into their own file
